deploy_ml/requirements.txt
pandas
matplotlib
wordcloud
streamlit
google-cloud-bigquery
matplotlib
seaborn
scikit-learn
nltk
db-dtypes